HashMap: quickfix infinite loop in Remove
Under some circumstances, when we try to remove a non-existent key
from a hash-map, we enter an infinite loop.

This is part of a deeper problem (those circumstances shouldn't
happen), but for now let's just detect the issue and return. This
is the same thing that other parts of the code does.