Practice safe formatting outside Windows too
Use C99 bounds-checking versions of the printf and scanf families of
functions.

Also, stop exporting our madness in externally visible #defines. While
we (mostly) control our use of the substituted functions, the Microsoft
versions are not completely equivalent to the standard C99 ones and care
must be taken. In String::Format we take a user-supplied format string
and pass it to either vsnprintf_s (MS-specific) or vsnprintf (C99),
which could become a problem.