Drop explicit BecomeVisible()
The point of this was to trigger OnDidStart from within the running
event loop. The way it's done, however races with
applicationDidBecomeActive, and makes it possible to leave the app in
the semi-useful state Visible.

OnDidStart will get called soon enough with DidBecomeActive, which is
just around the corner, anyway.