Text and keyboard events on iOS
InputEventHandler extended with getter for keyboard bounds (with default
implementation) and new OnClearTextInput event.

Experimented with an implementation of UIKeyInput protocol, but learned
we lose important key features, like repeat events for backspace
presses. The UITextInput protocol, on the other hand, would allow us to
build a full native text control but expects to merge deeply with the
control so isn't a nice fit for Xli's model.

In the end, followed SDL2's example with a hidden native text field that
redirects events into the InputEventHandler. Contrary to SDL2, however,
this only generates key events for Backspace and Return presses, other
input should be handled through the OnTextInput event.