Application lifecycle Background <> Foreground <> Interactive
The Visible state was renamed to Foreground, so as to lessen confusion
with Window actually being ready or displaying anything on the screen.

Active renamed to Interactive to reinforce that in this state we're
supposed to have full user engagement, or at least will be receiving
input events.

OnXxx events and their EmitOnXxx emitters were renamed accordingly. The
events around the Interactive state (formerly Active) were changed to
EnterX/ExitX format to fit in better.