Fix/tweak firing of OnDidStart
The transition from Starting to Background will now fire OnDidStart,
which was accidentally missing.

When transitioning to Terminating straight from Starting, we'll now
explicitly skip OnDidStart (previously missing), but also not enter
Background. This change assumes that a transition from Starting to
Terminating spells trouble in initialization code and we streamline the
exit process.