Remove android_native_glue. This had a lot of implications:
First off the reason: We added a choreographer, this meant having a
looper. This in turn blocks the thread so we had java drive the main
loop. This in turn meant that when we called into c++ we still needed
to pump the ndk messagequeue.

This seemed crazy as we were using less and less of the glue but paying
more of a cost for having it there, so we tear it out, change the
nativeactivity to an activity hook our own callbacks in On_JNILoad.

Now we have created an interesting fact, we are now in the java main
thread (which is also the UI thread). This meant we were having some
issues as there were two glcontexts in the same thread. For now I have
put makecurrent just before onupdateframe to deal with this.

Other implications come down to performance.
If we are doing everything on the main thread we could lag, if we lag
for too long we get killed. Without concurrency primitives in Uno we
cant encourage people to do the right thing, and even then we need to
make sure that Uno best practices are the same for ever platform (by
default)

I am going to add a feature to the android backend to optionally run in
a separate thread, this means if worst comes to the worst we can switch
it back on and we dont get held up nearly release time.