Don't cache SharedApp() beyond a single callback
In a future commit the returned value will change to a dummy handler
after OnTerminate has been fired.

Other than that, SharedApp() at the moment is an inline wrapper around a
static variable. Caching inside a single callback may be useful to let
the compiler know the value isn't changing, but beyond that it doesn't
buy much.