[iOS] Terminate application when UIApplicationMain returns
UIApplicationMain is not supposed to return. In the simulator, however,
we're able to catch exceptions from outside it and will call
OnUnhandledException from main.

This ensures we stop responding to events while showing the final error
message to the user. Going forward, we should better trap exceptional
conditions and ensure we're able to log or report the errors,
terminating the Application if appropriate.